Setup BigCommerce to have more than one backend per subsite.
This is necessary since the third_party_auth Django app on the platform relies on filtering the Email Provider configs by KEY_NAME=`backend_name` field which is the `name` for these social-core classes.

For any new subsite that requires BigCommerce Customer or Admin SSO we'll need to subclass the corresponding OAuth2 or Email backend.